+++ rs/sns_aggregator/src/types/slow.rs
@@ -2,7 +2,7 @@ use crate::types::ic_sns_governance::{GetMetadataResponse, ListNervousSystemFunc
 use crate::types::ic_sns_root::ListSnsCanistersResponse;
 use crate::types::ic_sns_swap::{DerivedState, GetStateResponse, Init, Params, Swap};
 use crate::types::ic_sns_wasm::DeployedSns;
-use crate::types::state::UpstreamData;
+use crate::types::upstream::UpstreamData;
 use crate::Icrc1Value;
 use candid::CandidType;
 use candid::Nat;
